Add files via upload
Modularized the creation of the instruction list & resolution of said instructions to their own methods. Implemented the skeleton of a "order_while_loop" method that works up to the decision to take or skip the branch of nested instructions. For assignment statements, if a left parenthese is read the user can skip lines and keep it as the same instruction until a right paren is matched, then the next skip line will end the example. ie. "a = ( 6 + \n 6) + 3 \n" is read as one instruction.