Add files via upload
Added string identification in createRawInstruction(), does not include the " or ' that encapsulate the string, only the contents. Added string functionality to getResultType, resolveMath, resolveCompare, resolveLogical, and resolveAssign. Updated convertTokenToValue to handle strings and variables that were declared to strings. Added runtime_error() method to allow the runtime to halt on error and display an error message; currently handles undeclared variables, incompatible types for an operation, incompatible operations for types, and invalid assignments. Changed the priority on "Not" statement to be the highest so it resolves before any other operation outside parentheses.